refactor(wiki): split category store impl into wiki_categories_store.go
Moves the inMemoryFactStore article_categories + category_parents methods and
the WikiIndex category passthroughs out of wiki_index.go (which had grown past
the 1500-LOC file-size budget) into a focused file. Pure code move — no
behavior change; the SQLite implementation already lives in wiki_index_sqlite.go.
